Music.Pitch

Description

Provides various representations of musical pitches and intervals.

Documentation

module Music.Pitch.Absolute

module Music.Pitch.Ambitus

module Music.Pitch.Equal

module Music.Pitch.Common

module Music.Pitch.Common.Names

data StaffLines

For notation systems using staff lines (i.e. CMN), this type represents staff number.

Rules:

Incrementing staff lines means moving upward.--

For staff systems with an odd number of lines, 0 is the middle staff.

For staff systems with an even number of lines, 0 is the staff below the middle space.

I.e. in CMN the staff lines are [-2..2], with negative numbers representing the lines below the middle line and positions numbers representing the lines above it.

data HalfSpaces

For notation systems using staff lines, this type represents the difference betwee two staff positions (i.e. one diatonic step in CMN).

Rules:

Incrementing half spaces means moving upward.

For staff systems with an odd number of lines, 0 is the middle staff.

For staff systems with an even number of lines, 0 is the middle space.

I.e. in CMN, the non-ledger positions are [-5..5] with the even numbers representing lines and the odd representing spaces and with positive numbers representing positions above the middle line and negative numbers representing positions below.

module Music.Pitch.Clef

module Music.Pitch.Intonation