Music.Pitch.Ambitus

Description

Musical ambitus, or pitch ranges.

Documentation

newtype Ambitus a

An ambitus is a closed interval (in the mathematical sense).

Also known as range or tessitura, this type can be used to restrict the range of a melody, chord or other pitch container.

It is also used in music-parts to represent the range of instruments.

Constructors

Ambitus

Fields

getAmbitus :: Interval a

ambitus :: (Num a, Ord a) => Iso' (a, a) (Ambitus a)