Music.Pitch.Common.Quality

Description

Common interval quality.

Quality

data Quality

The quality component of an interval (minor, major, augmented). Generalized from single/double augmented/diminished to any number of steps.

Constructors

Major

Minor

Perfect

Augmented Integer

Diminished Integer

qualityTypes :: Quality -> [QualityType]

Return all possible quality types for a given quality.

class HasQuality a where

Types of value that has an interval quality (mainly Interval and Quality itself).

Methods

quality :: a -> Quality

invertQuality :: Quality -> Quality

Invert a quality.

Perfect is unaffected, major becomes minor and vice versa, augmented becomes diminished and vice versa.

isPerfect :: HasQuality a => a -> Bool

Returns whether the given quality is perfect.

isMajor :: HasQuality a => a -> Bool

Returns whether the given quality is major.

isMinor :: HasQuality a => a -> Bool

Returns whether the given quality is minor.

isAugmented :: HasQuality a => a -> Bool

Returns whether the given quality is augmented (including double augmented etc).

isDiminished :: HasQuality a => a -> Bool

Returns whether the given quality is diminished (including double diminished etc).

Quality type

data QualityType

The alteration implied by a quality is dependent on whether it is attached to a major/minor vs. a perfect-style number. This type represents the two possibilities.

Constructors

PerfectType

MajorMinorType

expectedQualityType :: Number -> QualityType

The quality type expected for a given number, i.e. perfect for unisons, fourths, and fifths and major/minor for everything else.

Quality to alteration

data Direction

Constructors

Upward

Downward

qualityToAlteration :: Direction -> QualityType -> Quality -> Maybe ChromaticSteps

Return the alteration in implied by the given quality to a number of a given quality type.

qualityToDiff :: Bool -> QualityType -> Quality -> ChromaticSteps

Deprecated: Use qualityToAlteration