Music.Pitch.Scale

Description

Scales and chords.

Modes and scales

data Mode a

A mode is a list of intervals and a characteristic repeating interval.

modeFromSteps :: [Diff a] -> Diff a -> Mode a

[Interval] -> Interval -> Mode Pitch

modeIntervals :: Lens' (Mode a) [Diff a]

Lens' (Mode Pitch) [Interval]

modeRepeat :: Lens' (Mode a) (Diff a)

Lens' (Mode Pitch) Interval

data Scale a

A scale is a mode with a specified tonic.

scaleTonic :: Lens' (Scale a) a

Lens' (Scale Pitch) Pitch

scaleMode :: Lens' (Scale a) (Mode a)

Lens' (Scale Pitch) (Mode Pitch)

leadingInterval :: AffineSpace a => Mode a -> Diff a

>>> leadingInterval majorScale
m2
>>> leadingInterval harmonicMinorScale
m2
>>> leadingInterval pureMinorScale
_M2

invertMode :: AffineSpace a => Int -> Mode a -> Mode a

modeToScale :: AffineSpace a => a -> Mode a -> Scale a

scaleToList :: AffineSpace a => Scale a -> [a]

Chord types and chords

data Function a

functionFromSteps :: [Diff a] -> Diff a -> Function a

[Interval] -> Interval -> Function Pitch

functionIntervals :: Lens' (Function a) [Diff a]

Lens' (Function Pitch) [Interval]

functionRepeat :: Lens' (Function a) (Diff a)

Lens' (Function Pitch) Interval

data Chord a

chordTonic :: Lens' (Chord a) a

Lens' (Chord Pitch) Pitch

chordFunction :: Lens' (Chord a) (Function a)

Lens' (Chord Pitch) (Function Pitch)

complementInterval :: AffineSpace a => Function a -> Diff a

>>> complementInterval majorTriad
_P4
>>> complementInterval minorTriad
_P4
>>> complementInterval majorMinorSeventhChord
_M2

Lens' (Function Pitch) Interval

invertChord :: AffineSpace a => Int -> Function a -> Function a

functionToChord :: AffineSpace a => a -> Function a -> Chord a

chordToList :: AffineSpace a => Chord a -> [a]

Returns a single inversion of the given chord (no repeats!).

Common modes

Classical modes

majorScale :: Mode Pitch

pureMinorScale :: Mode Pitch

harmonicMinorScale :: Mode Pitch

melodicMinorScaleUp :: Mode Pitch

Church modes

aeolian :: Mode Pitch

locrian :: Mode Pitch

ionian :: Mode Pitch

dorian :: Mode Pitch

phrygian :: Mode Pitch

lydian :: Mode Pitch

mixolydian :: Mode Pitch

majorPentaTonic :: Mode Pitch

minorPentaTonic :: Mode Pitch

bluesMinor :: Mode Pitch

bluesMajor :: Mode Pitch

bebopScale :: Mode Pitch

Miscellaneous modes

wholeTone :: Mode Pitch

octatonic :: Mode Pitch

Modes of limited transposition

firstMode :: Mode Pitch

secondMode :: Mode Pitch

thirdMode :: Mode Pitch

fourthMode :: Mode Pitch

fifthMode :: Mode Pitch

sixthMode :: Mode Pitch

seventhMode :: Mode Pitch

Common chords

majorTriad :: Function Pitch

minorTriad :: Function Pitch

augmentedChord :: Function Pitch

diminishedChord :: Function Pitch

halfDiminishedChord :: Function Pitch

majorMinorSeventhChord :: Function Pitch

majorMajorSeventhChord :: Function Pitch