Music.Pitch.Absolute

Description

Absolute pitch representation.

The canonical pitch representation is frequency in Hertz (Hz).

For conversion, see HasFrequency.

Absolute pitch representation

newtype Hertz

Absolute frequency in Hertz.

Constructors

Hertz

Fields

getHertz :: Double

data Cents

Number of pure octaves.

Cents are a logarithmic representation of frequency such that

f * (2/1) = frequency (cents f + 1200)

data Fifths

Number of pure fifths.

Fifths are a logarithmic representation of frequency.

f * (3/2) = frequency (fifths f + 1)

HasFrequency class

class HasFrequency a where

Methods

frequency :: a -> Hertz

fifths :: HasFrequency a => a -> Fifths

cents :: HasFrequency a => a -> Cents