Music.Pitch.Ambitus

Description

Pitch range or ambitus.

Documentation

data Ambitus a

An ambitus is (mathematical) interval.

Also known as range or tessitura, this type can be used to restrict the range instruments, chords, melodies etc.

ambitus :: (Num a, Ord a) => Iso (a, a) (b, b) (Ambitus a) (Ambitus b)

mapAmbitus :: (Ord b, Num b) => (a -> b) -> Ambitus a -> Ambitus b

Not a true functor for similar reasons as sets.

ambitusHighest :: (Num a, Ord a) => Ambitus a -> a

ambitusLowest :: (Num a, Ord a) => Ambitus a -> a

ambitusInterval :: (Num a, Ord a, AffineSpace a) => Ambitus a -> Diff a

Returns a postive interval (or _P1 for empty ambitus)