Music.Pitch.Common.Quality

Quality

data Quality

Interval quality is either perfect, major, minor, augmented, and diminished. This representation allows for an arbitrary number of augmentations or diminutions, so augmented is represented by Augmented 1, doubly augmented by Augmented 2 and so on.

The quality of a compound interval is the quality of the simple interval on which it is based.

Constructors

Major

Minor

Perfect

Augmented Integer

Diminished Integer

class HasQuality a where

Types of value that has an interval quality (mainly Interval and Quality itself).

Methods

quality :: a -> Quality

invertQuality :: Quality -> Quality

Invert a quality.

Perfect is unaffected, major becomes minor and vice versa, augmented becomes diminished and vice versa.

isPerfect :: HasQuality a => a -> Bool

Returns whether the given quality is perfect.

isMajor :: HasQuality a => a -> Bool

Returns whether the given quality is major.

isMinor :: HasQuality a => a -> Bool

Returns whether the given quality is minor.

isAugmented :: HasQuality a => a -> Bool

Returns whether the given quality is augmented (including double augmented etc).

isDiminished :: HasQuality a => a -> Bool

Returns whether the given quality is diminished (including double diminished etc).

Quality type

data QualityType

Constructors

PerfectType

MajorMinorType

expectedQualityType :: HasNumber a => a -> QualityType

qualityTypes :: Quality -> [QualityType]

qualityToDiff :: Bool -> QualityType -> Quality -> ChromaticSteps